use fabsf instead of fabs for floats

+++ library/tinyspline.c
@@ -692,8 +692,8 @@ int ts_fequals(const float x, const float y)
     if (fabs(x-y) <= FLT_MAX_ABS_ERROR) {
         return 1;
     } else {
-        const float r = fabs(x) > fabs(y) ? 
-            fabs((x-y) / x) : fabs((x-y) / y);
+        const float r = fabsf(x) > fabsf(y) ?
+            fabsf((x-y) / x) : fabsf((x-y) / y);
         return r <= FLT_MAX_REL_ERROR;
     }
 }
